refactor(report): extract re-rendering helpers to rerender.py
Move MetricStub, is_session_data_stripped, and
metric_stubs_from_metadata from cli.py to src/raki/report/rerender.py.
Reduces cli.py by ~80 lines and prepares helpers for reuse.

Refs #82

Assigned-by: decko
